# server/admin/docs/swagger.yaml
basePath: /api/v1/admin
definitions:
  admin.ErrorResponse:
    properties:
      error:
        example: resource not found
        type: string
    type: object
...
  store.CronJob:
    properties:
      agentId:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      prompt:
        type: string
      schedule:
        type: string
    type: object
  store.Device:
    properties:
      allowedAgents:
        items:
          type: string
        type: array
      defaultAgent:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      token:
        type: string
    type: object
  store.MCPServer:
    properties:
      args:
        items:
          type: string
        type: array
      command:
        type: string
      endpoint:
        type: string
      env:
        additionalProperties:
          type: string
        type: object
      headers:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      systemPrompt:
        type: string
      type:
        type: string
      workDir:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Administration API for managing Magec multi-agent voice assistant resources.
  title: Magec Admin API
  version: "1.0"
paths:
...
  /crons:
    get:
      description: Returns all configured scheduled tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.CronJob'
            type: array
      summary: List cron jobs
      tags:
      - crons
    post:
      consumes:
      - application/json
      description: Creates a new scheduled task that sends a prompt to an agent
      parameters:
      - description: Cron job definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.CronJob'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.CronJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Create cron job
      tags:
      - crons
  /crons/{name}:
    delete:
      description: Deletes a cron job by name
      parameters:
      - description: Cron job name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Delete cron job
      tags:
      - crons
    get:
      description: Returns a cron job by its unique name
      parameters:
      - description: Cron job name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.CronJob'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Get cron job
      tags:
      - crons
    put:
      consumes:
      - application/json
      description: Updates a cron job by name
      parameters:
      - description: Cron job name
        in: path
        name: name
        required: true
        type: string
      - description: Cron job definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.CronJob'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.CronJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Update cron job
      tags:
      - crons
  /devices:
    get:
      description: Returns all configured access point devices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Device'
            type: array
      summary: List devices
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Creates a new device. A unique auth token is generated automatically.
      parameters:
      - description: Device definition (token is auto-generated)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.Device'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Create device
      tags:
      - devices
  /devices/{name}:
    delete:
      description: Deletes a device by name, revoking its access token
      parameters:
      - description: Device name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Delete device
      tags:
      - devices
    get:
      description: Returns a device by its unique name
      parameters:
      - description: Device name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Device'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Get device
      tags:
      - devices
    put:
      consumes:
      - application/json
      description: Updates a device by name. Token is preserved.
      parameters:
      - description: Device name
        in: path
        name: name
        required: true
        type: string
      - description: Device definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Update device
      tags:
      - devices
  /devices/{name}/regenerate-token:
    post:
      description: Generates a new authentication token for a device, invalidating
        the previous one
      parameters:
      - description: Device name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Device'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Regenerate device token
      tags:
      - devices
  /mcps:
    get:
      description: Returns all configured MCP tool servers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.MCPServer'
            type: array
      summary: List MCP servers
      tags:
      - mcps
    post:
      consumes:
      - application/json
      description: Creates a new MCP tool server
      parameters:
      - description: MCP server definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.MCPServer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.MCPServer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Create MCP server
      tags:
      - mcps
  /mcps/{name}:
    delete:
      description: Deletes an MCP server by name
      parameters:
      - description: MCP server name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Delete MCP server
      tags:
      - mcps
    get:
      description: Returns an MCP server by its unique name
      parameters:
      - description: MCP server name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.MCPServer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Get MCP server
      tags:
      - mcps
    put:
      consumes:
      - application/json
      description: Updates an MCP server by name
      parameters:
      - description: MCP server name
        in: path
        name: name
        required: true
        type: string
      - description: MCP server definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.MCPServer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.MCPServer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Update MCP server
      tags:
      - mcps
...
schemes:
- http
swagger: "2.0"